record: Fix use of std::min
Update #3665.

+++ trace/record/record-client-base.cc
@@ -40,6 +40,7 @@
 #include <fcntl.h>
 #include <sys/stat.h>
 
+#include <algorithm>
 #include <cassert>
 #include <cstring>
 
@@ -108,7 +109,7 @@ void Client::Run() {
 
   while (stop_ == 0 && todo > 0) {
     int buf[8192];
-    size_t m = std::min(sizeof(buf), todo);
+    size_t m = std::min(static_cast<uint64_t>(sizeof(buf)), todo);
     ssize_t n = input_.Read(buf, m);
     if (n <= 0) {
       break;
